+++ startup/GafferScene/standardOptions.py
@@ -280,6 +280,7 @@ Gaffer.Metadata.registerValues( {
 		InteractiveRender nodes.
 		""",
 		"label", "Default Renderer",
+		"columnLayout:label", "Renderer",
 		"layout:section", "Renderer",
 
 		"plugValueWidget:type", "GafferUI.PresetsPlugValueWidget",
@@ -377,6 +378,7 @@ Gaffer.Metadata.registerValues( {
 		# for others. We should look into whether shipping a standard font with Gaffer would reduce
 		# this ambiguity.
 		"label", "Camera Inclusions",
+		"columnLayout:label", "Camera Inclusions / Catchers",
 		"layout:section", "Visibility Set",
 
 		"plugValueWidget:type", "GafferSceneUI.SetExpressionPlugValueWidget",
@@ -402,6 +404,7 @@ Gaffer.Metadata.registerValues( {
 		in the scene take precedence over this option.
 		""",
 		"label", "Camera Exclusions",
+		"columnLayout:label", "Camera Exclusions / Casters",
 		"layout:section", "Visibility Set",
 
 		"plugValueWidget:type", "GafferSceneUI.SetExpressionPlugValueWidget",
@@ -459,6 +462,7 @@ Gaffer.Metadata.registerValues( {
 		node with appropriate filters.
 		""",
 		"label", "Transform",
+		"columnLayout:label", "Transform Blur",
 		"layout:section", "Motion Blur",
 
 	],
@@ -475,6 +479,7 @@ Gaffer.Metadata.registerValues( {
 		node with appropriate filters.
 		""",
 		"label", "Deformation",
+		"columnLayout:label", "Deformation Blur",
 		"layout:section", "Motion Blur",
 
 	],