+++ Changes.md
@@ -1,6 +1,15 @@
 1.6.x.x (relative to 1.6.2.1)
 =======
 
+Improvements
+------------
+
+- RenderPassEditor :
+  - Added "Favourites" section. Columns from any other section can be marked for display in this section by right-clicking on their header and selecting "Favourite" from the menu. Columns can be reordered within the Favourites section by dragging their header.
+  - Any standard or renderer-specific option can be added as a favourite by clicking on the `+` column header and choosing it from the menu, or any option can be added via drag & drop from the SceneInspector.
+  - A default set of favourite columns can be configured by registering metadata containing a list of option names in a startup file. `Gaffer.Metadata.registerValue( GafferSceneUI.RenderPassEditor.Settings, "favouriteColumns", "userDefault", IECore.StringVectorData( [ "option:render:cameraInclusions", "option:render:matteInclusions" ] ) )`
+- Spreadsheet : Improved visual feedback when reordering columns via drag & drop.
+
 Fixes
 -----
 
@@ -10,6 +19,7 @@ API
 ---
 
 - Metadata : The `registerNode()` function now accepts dictionaries containing plug metadata. This should be preferred to the previous list-based values.
+- SceneInspector : Added `draggedAttributeData()`, `draggedOptionData()` and `draggedParameterData()` static methods.
 
 1.6.2.1 (relative to 1.6.2.0)
 =======